requirements.txt
# L9 Core Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Pydantic Settings (separate package in v2)
pydantic-settings>=2.1.0

# Database
psycopg[binary]>=3.1.14
sqlalchemy>=2.0.23
asyncpg>=0.29.0
pgvector>=0.2.4

# Redis
redis>=5.0.0

# Neo4j
neo4j>=5.14.0

# HTTP Client
httpx>=0.26.0

# OpenAI / LLM
openai>=1.10.0

# LangGraph / LangChain
langgraph>=0.0.40
langchain-core>=0.1.20

# Logging
structlog>=24.1.0

# Async utilities
anyio>=4.2.0
aiofiles>=23.2.0

# Typing extensions
typing-extensions>=4.9.0

# Slack
slack-sdk>=3.26.0

# Twilio (SMS/Voice)
twilio>=8.10.0

# Templates
Jinja2>=3.1.2

# YAML
PyYAML>=6.0.1

# Cryptography (signature verification)
cryptography>=41.0.0

# Symbolic Computation (Quantum AI Factory)
sympy>=1.12
numpy>=1.24.0

# Browser Automation (Mac Agent)
playwright>=1.40.0

# Testing (optional, for dev)
pytest>=7.4.0
pytest-asyncio>=0.23.0
python-multipart
